feat: One-time password for operations
Upon DID creation, the create operation payload must include:
-- The hash of a one-time password (OTP) for the next recovery operation.
-- The hash of a one-time password (OTP) for the next update operation.

The DID owner must reproduce and present the correct OTP in the subsequent operation for the operation to be considered valid. In addition, each subsequent operation must also include the hash of the new OTP(s) for the next operation. This scheme enables efficient dismissal of counterfeit operations without needing to evaluate signatures.

Closes #90

Signed-off-by: Sandra Vrtikapa <sandra.vrtikapa@securekey.com>