fix: Always start batch writer timer
This commit ensures that the batch writer timer is always running, even if no operations were added. Previously, the batch writer timer was only started if an operation had been added locally. In the case where the operation queue is distributed (e.g. message broker) then an operation may have been added to Server1, published to a distributed queue and delivered to Server2, but the timer was not started on Server2 and so the message was not processed (until another operation is added to Server2).

closes #590

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>