fix: Adding new operations don't wait for batch to be cut
Previously, adding a new operation may need to wait for the batch to be cut if the buffer of the Go channel has been maxed out. This commit changes the behaviour such that adding a new operation does not wait on the Go channel. Instead, the queue is monitored in the background (by default, every 1 second) to see if the maximum batch size has been reached. If so, the batch is cut.

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>