feat: Add alternate sources for CAS files during Observer processing
When the Observer processes a batch, it reads various Sidetree files from CAS, starting with the core index file. If a file is not found, then one or more alternate sources are checked. This is useful in recovery, i.e. when data at the original source is lost but may be recovered by reading it at an alternate source.

closes #637

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>