fix: Fix race condition in resolve
Fixes a race condition which may occur when published operations are first retrieved, not found, and then unpublished operations are deleted before they are retrieved. In this case the DID resolve results in a 404 "not found" error.

This commit reverses the order of retrieval, i.e. first retrieve the unpublished operations and then the published ones.

closes #642

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>