chore: Remove "updated" for deactivated document
Since Sidetree resolution is processing full operations first (create, recover, deactivate) having "updated" in the resolution result for deactivate document will lead to inconsistent behaviour. If there was a 'recover' operation in the chain it will return "updated" otherwise it will be empty. Also in this case "updated" means the anchoring time of last recover and not necessarily the last update.

Closes #668

Signed-off-by: Sandra Vrtikapa <sandra.vrtikapa@securekey.com>